+++ src/components.d.ts
@@ -5,8 +5,8 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { CustomToastType, IComponentToast, ISimpleToast } from "./components/toasts-list/components/transaction-toast/transaction-toast.type";
 import { IconDefinition } from "@fortawesome/free-solid-svg-icons";
+import { CustomToastType, IComponentToast, ISimpleToast } from "./components/toasts-list/components/transaction-toast/transaction-toast.type";
 import { LocalJSX as JSX, VNode } from "@stencil/core";
 import { ILedgerConnectModalData } from "./components/ledger-connect-modal/ledger-connect-modal.types";
 import { IEventBus } from "./utils/EventBus";
@@ -14,9 +14,10 @@ import { IPendingTransactionsModalData } from "./components/pending-transactions
 import { ISignTransactionsModalData } from "./components/sign-transactions-modal/sign-transactions-modal.types";
 import { CustomToastType as CustomToastType1, IToastDataState, ITransaction, ITransactionProgressState, ITransactionToast } from "./components/toasts-list/components/transaction-toast/transaction-toast.type";
 import { ITransactionAccount, ITransactionIconInfo, ITransactionsTableRow } from "./components/transactions-table/transactions-table.type";
+import { ITransactionValue } from "./components/transactions-table/transactions-table.type";
 import { IWalletConnectModalData } from "./components/wallet-connect-modal/wallet-connect-modal.types";
-export { CustomToastType, IComponentToast, ISimpleToast } from "./components/toasts-list/components/transaction-toast/transaction-toast.type";
 export { IconDefinition } from "@fortawesome/free-solid-svg-icons";
+export { CustomToastType, IComponentToast, ISimpleToast } from "./components/toasts-list/components/transaction-toast/transaction-toast.type";
 export { LocalJSX as JSX, VNode } from "@stencil/core";
 export { ILedgerConnectModalData } from "./components/ledger-connect-modal/ledger-connect-modal.types";
 export { IEventBus } from "./utils/EventBus";
@@ -24,6 +25,7 @@ export { IPendingTransactionsModalData } from "./components/pending-transactions
 export { ISignTransactionsModalData } from "./components/sign-transactions-modal/sign-transactions-modal.types";
 export { CustomToastType as CustomToastType1, IToastDataState, ITransaction, ITransactionProgressState, ITransactionToast } from "./components/toasts-list/components/transaction-toast/transaction-toast.type";
 export { ITransactionAccount, ITransactionIconInfo, ITransactionsTableRow } from "./components/transactions-table/transactions-table.type";
+export { ITransactionValue } from "./components/transactions-table/transactions-table.type";
 export { IWalletConnectModalData } from "./components/wallet-connect-modal/wallet-connect-modal.types";
 export namespace Components {
     interface BalanceComponent {
@@ -31,6 +33,12 @@ export namespace Components {
         "header"?: string;
         "ticker": string;
         "usdValue"?: string;
+    }
+    interface CopyButton {
+        "class"?: string;
+        "copyIcon": IconDefinition;
+        "successIcon": IconDefinition;
+        "text": string;
     }
     interface CustomToast {
         "toast": IComponentToast;
@@ -42,7 +50,7 @@ export namespace Components {
         "link": string;
         "text"?: string;
     }
-    interface FontawesomeIcon {
+    interface FaIcon {
         "class"?: string;
         "description"?: string;
         "icon": IconDefinition;
@@ -125,6 +133,10 @@ export namespace Components {
         "class"?: string;
         "tooltip"?: string;
     }
+    interface TransactionDirectionBadge {
+        "class"?: string;
+        "direction": string;
+    }
     interface TransactionHash {
         "class"?: string;
         "transaction": ITransactionsTableRow;
@@ -139,6 +151,10 @@ export namespace Components {
         "method": string;
     }
     interface TransactionRow {
+        "class"?: string;
+        "transaction": ITransactionsTableRow;
+    }
+    interface TransactionShards {
         "class"?: string;
         "transaction": ITransactionsTableRow;
     }
@@ -173,9 +189,18 @@ export namespace Components {
         "wrapperClass": string;
         "wrapperId"?: string;
     }
+    interface TransactionValue {
+        "class"?: string;
+        "value": ITransactionValue;
+    }
     interface TransactionsTable {
         "class"?: string;
         "data": string;
+    }
+    interface TrimText {
+        "class"?: string;
+        "dataTestId": string;
+        "text": string;
     }
     interface WalletConnectModal {
         "data": IWalletConnectModalData;
@@ -213,6 +238,12 @@ declare global {
         prototype: HTMLBalanceComponentElement;
         new (): HTMLBalanceComponentElement;
     };
+    interface HTMLCopyButtonElement extends Components.CopyButton, HTMLStencilElement {
+    }
+    var HTMLCopyButtonElement: {
+        prototype: HTMLCopyButtonElement;
+        new (): HTMLCopyButtonElement;
+    };
     interface HTMLCustomToastElementEventMap {
         "handleDeleteToast": string;
     }
@@ -236,11 +267,11 @@ declare global {
         prototype: HTMLExplorerLinkElement;
         new (): HTMLExplorerLinkElement;
     };
-    interface HTMLFontawesomeIconElement extends Components.FontawesomeIcon, HTMLStencilElement {
+    interface HTMLFaIconElement extends Components.FaIcon, HTMLStencilElement {
     }
-    var HTMLFontawesomeIconElement: {
-        prototype: HTMLFontawesomeIconElement;
-        new (): HTMLFontawesomeIconElement;
+    var HTMLFaIconElement: {
+        prototype: HTMLFaIconElement;
+        new (): HTMLFaIconElement;
     };
     interface HTMLFormatAmountElement extends Components.FormatAmount, HTMLStencilElement {
     }
@@ -371,6 +402,12 @@ declare global {
         prototype: HTMLTransactionAgeElement;
         new (): HTMLTransactionAgeElement;
     };
+    interface HTMLTransactionDirectionBadgeElement extends Components.TransactionDirectionBadge, HTMLStencilElement {
+    }
+    var HTMLTransactionDirectionBadgeElement: {
+        prototype: HTMLTransactionDirectionBadgeElement;
+        new (): HTMLTransactionDirectionBadgeElement;
+    };
     interface HTMLTransactionHashElement extends Components.TransactionHash, HTMLStencilElement {
     }
     var HTMLTransactionHashElement: {
@@ -394,6 +431,12 @@ declare global {
     var HTMLTransactionRowElement: {
         prototype: HTMLTransactionRowElement;
         new (): HTMLTransactionRowElement;
+    };
+    interface HTMLTransactionShardsElement extends Components.TransactionShards, HTMLStencilElement {
+    }
+    var HTMLTransactionShardsElement: {
+        prototype: HTMLTransactionShardsElement;
+        new (): HTMLTransactionShardsElement;
     };
     interface HTMLTransactionToastElementEventMap {
         "handleDeleteToast": string;
@@ -453,11 +496,23 @@ declare global {
         prototype: HTMLTransactionToastWrapperElement;
         new (): HTMLTransactionToastWrapperElement;
     };
+    interface HTMLTransactionValueElement extends Components.TransactionValue, HTMLStencilElement {
+    }
+    var HTMLTransactionValueElement: {
+        prototype: HTMLTransactionValueElement;
+        new (): HTMLTransactionValueElement;
+    };
     interface HTMLTransactionsTableElement extends Components.TransactionsTable, HTMLStencilElement {
     }
     var HTMLTransactionsTableElement: {
         prototype: HTMLTransactionsTableElement;
         new (): HTMLTransactionsTableElement;
+    };
+    interface HTMLTrimTextElement extends Components.TrimText, HTMLStencilElement {
+    }
+    var HTMLTrimTextElement: {
+        prototype: HTMLTrimTextElement;
+        new (): HTMLTrimTextElement;
     };
     interface HTMLWalletConnectModalElement extends Components.WalletConnectModal, HTMLStencilElement {
     }
@@ -467,9 +522,10 @@ declare global {
     };
     interface HTMLElementTagNameMap {
         "balance-component": HTMLBalanceComponentElement;
+        "copy-button": HTMLCopyButtonElement;
         "custom-toast": HTMLCustomToastElement;
         "explorer-link": HTMLExplorerLinkElement;
-        "fontawesome-icon": HTMLFontawesomeIconElement;
+        "fa-icon": HTMLFaIconElement;
         "format-amount": HTMLFormatAmountElement;
         "fungible-component": HTMLFungibleComponentElement;
         "generic-modal": HTMLGenericModalElement;
@@ -486,17 +542,21 @@ declare global {
         "transaction-account": HTMLTransactionAccountElement;
         "transaction-account-name": HTMLTransactionAccountNameElement;
         "transaction-age": HTMLTransactionAgeElement;
+        "transaction-direction-badge": HTMLTransactionDirectionBadgeElement;
         "transaction-hash": HTMLTransactionHashElement;
         "transaction-icon": HTMLTransactionIconElement;
         "transaction-method": HTMLTransactionMethodElement;
         "transaction-row": HTMLTransactionRowElement;
+        "transaction-shards": HTMLTransactionShardsElement;
         "transaction-toast": HTMLTransactionToastElement;
         "transaction-toast-content": HTMLTransactionToastContentElement;
         "transaction-toast-details": HTMLTransactionToastDetailsElement;
         "transaction-toast-details-body": HTMLTransactionToastDetailsBodyElement;
         "transaction-toast-progress": HTMLTransactionToastProgressElement;
         "transaction-toast-wrapper": HTMLTransactionToastWrapperElement;
+        "transaction-value": HTMLTransactionValueElement;
         "transactions-table": HTMLTransactionsTableElement;
+        "trim-text": HTMLTrimTextElement;
         "wallet-connect-modal": HTMLWalletConnectModalElement;
     }
 }
@@ -506,6 +566,12 @@ declare namespace LocalJSX {
         "header"?: string;
         "ticker"?: string;
         "usdValue"?: string;
+    }
+    interface CopyButton {
+        "class"?: string;
+        "copyIcon"?: IconDefinition;
+        "successIcon"?: IconDefinition;
+        "text"?: string;
     }
     interface CustomToast {
         "onHandleDeleteToast"?: (event: CustomToastCustomEvent<string>) => void;
@@ -518,7 +584,7 @@ declare namespace LocalJSX {
         "link"?: string;
         "text"?: string;
     }
-    interface FontawesomeIcon {
+    interface FaIcon {
         "class"?: string;
         "description"?: string;
         "icon"?: IconDefinition;
@@ -600,6 +666,10 @@ declare namespace LocalJSX {
         "class"?: string;
         "tooltip"?: string;
     }
+    interface TransactionDirectionBadge {
+        "class"?: string;
+        "direction"?: string;
+    }
     interface TransactionHash {
         "class"?: string;
         "transaction"?: ITransactionsTableRow;
@@ -614,6 +684,10 @@ declare namespace LocalJSX {
         "method"?: string;
     }
     interface TransactionRow {
+        "class"?: string;
+        "transaction"?: ITransactionsTableRow;
+    }
+    interface TransactionShards {
         "class"?: string;
         "transaction"?: ITransactionsTableRow;
     }
@@ -650,18 +724,28 @@ declare namespace LocalJSX {
         "wrapperClass"?: string;
         "wrapperId"?: string;
     }
+    interface TransactionValue {
+        "class"?: string;
+        "value"?: ITransactionValue;
+    }
     interface TransactionsTable {
         "class"?: string;
         "data"?: string;
+    }
+    interface TrimText {
+        "class"?: string;
+        "dataTestId"?: string;
+        "text"?: string;
     }
     interface WalletConnectModal {
         "data"?: IWalletConnectModalData;
     }
     interface IntrinsicElements {
         "balance-component": BalanceComponent;
+        "copy-button": CopyButton;
         "custom-toast": CustomToast;
         "explorer-link": ExplorerLink;
-        "fontawesome-icon": FontawesomeIcon;
+        "fa-icon": FaIcon;
         "format-amount": FormatAmount;
         "fungible-component": FungibleComponent;
         "generic-modal": GenericModal;
@@ -678,17 +762,21 @@ declare namespace LocalJSX {
         "transaction-account": TransactionAccount;
         "transaction-account-name": TransactionAccountName;
         "transaction-age": TransactionAge;
+        "transaction-direction-badge": TransactionDirectionBadge;
         "transaction-hash": TransactionHash;
         "transaction-icon": TransactionIcon;
         "transaction-method": TransactionMethod;
         "transaction-row": TransactionRow;
+        "transaction-shards": TransactionShards;
         "transaction-toast": TransactionToast;
         "transaction-toast-content": TransactionToastContent;
         "transaction-toast-details": TransactionToastDetails;
         "transaction-toast-details-body": TransactionToastDetailsBody;
         "transaction-toast-progress": TransactionToastProgress;
         "transaction-toast-wrapper": TransactionToastWrapper;
+        "transaction-value": TransactionValue;
         "transactions-table": TransactionsTable;
+        "trim-text": TrimText;
         "wallet-connect-modal": WalletConnectModal;
     }
 }
@@ -697,9 +785,10 @@ declare module "@stencil/core" {
     export namespace JSX {
         interface IntrinsicElements {
             "balance-component": LocalJSX.BalanceComponent & JSXBase.HTMLAttributes<HTMLBalanceComponentElement>;
+            "copy-button": LocalJSX.CopyButton & JSXBase.HTMLAttributes<HTMLCopyButtonElement>;
             "custom-toast": LocalJSX.CustomToast & JSXBase.HTMLAttributes<HTMLCustomToastElement>;
             "explorer-link": LocalJSX.ExplorerLink & JSXBase.HTMLAttributes<HTMLExplorerLinkElement>;
-            "fontawesome-icon": LocalJSX.FontawesomeIcon & JSXBase.HTMLAttributes<HTMLFontawesomeIconElement>;
+            "fa-icon": LocalJSX.FaIcon & JSXBase.HTMLAttributes<HTMLFaIconElement>;
             "format-amount": LocalJSX.FormatAmount & JSXBase.HTMLAttributes<HTMLFormatAmountElement>;
             "fungible-component": LocalJSX.FungibleComponent & JSXBase.HTMLAttributes<HTMLFungibleComponentElement>;
             "generic-modal": LocalJSX.GenericModal & JSXBase.HTMLAttributes<HTMLGenericModalElement>;
@@ -716,17 +805,21 @@ declare module "@stencil/core" {
             "transaction-account": LocalJSX.TransactionAccount & JSXBase.HTMLAttributes<HTMLTransactionAccountElement>;
             "transaction-account-name": LocalJSX.TransactionAccountName & JSXBase.HTMLAttributes<HTMLTransactionAccountNameElement>;
             "transaction-age": LocalJSX.TransactionAge & JSXBase.HTMLAttributes<HTMLTransactionAgeElement>;
+            "transaction-direction-badge": LocalJSX.TransactionDirectionBadge & JSXBase.HTMLAttributes<HTMLTransactionDirectionBadgeElement>;
             "transaction-hash": LocalJSX.TransactionHash & JSXBase.HTMLAttributes<HTMLTransactionHashElement>;
             "transaction-icon": LocalJSX.TransactionIcon & JSXBase.HTMLAttributes<HTMLTransactionIconElement>;
             "transaction-method": LocalJSX.TransactionMethod & JSXBase.HTMLAttributes<HTMLTransactionMethodElement>;
             "transaction-row": LocalJSX.TransactionRow & JSXBase.HTMLAttributes<HTMLTransactionRowElement>;
+            "transaction-shards": LocalJSX.TransactionShards & JSXBase.HTMLAttributes<HTMLTransactionShardsElement>;
             "transaction-toast": LocalJSX.TransactionToast & JSXBase.HTMLAttributes<HTMLTransactionToastElement>;
             "transaction-toast-content": LocalJSX.TransactionToastContent & JSXBase.HTMLAttributes<HTMLTransactionToastContentElement>;
             "transaction-toast-details": LocalJSX.TransactionToastDetails & JSXBase.HTMLAttributes<HTMLTransactionToastDetailsElement>;
             "transaction-toast-details-body": LocalJSX.TransactionToastDetailsBody & JSXBase.HTMLAttributes<HTMLTransactionToastDetailsBodyElement>;
             "transaction-toast-progress": LocalJSX.TransactionToastProgress & JSXBase.HTMLAttributes<HTMLTransactionToastProgressElement>;
             "transaction-toast-wrapper": LocalJSX.TransactionToastWrapper & JSXBase.HTMLAttributes<HTMLTransactionToastWrapperElement>;
+            "transaction-value": LocalJSX.TransactionValue & JSXBase.HTMLAttributes<HTMLTransactionValueElement>;
             "transactions-table": LocalJSX.TransactionsTable & JSXBase.HTMLAttributes<HTMLTransactionsTableElement>;
+            "trim-text": LocalJSX.TrimText & JSXBase.HTMLAttributes<HTMLTrimTextElement>;
             "wallet-connect-modal": LocalJSX.WalletConnectModal & JSXBase.HTMLAttributes<HTMLWalletConnectModalElement>;
         }
     }